Tests: Ensure compatibility with rack_test driver
The visit calls in the system tests are using the url helpers,
such as brands_url. These don't play nicely with Capybara and
the rack_test driver. Replace these *_url helper calls with *_path.

In addition, the deletion system tests use the accept_confirm
helper to simulate clicking on the 'ok' button in a deletion
dialog that is driven by javascript. This isn't necessary with the
rack_test config. Remove the use of accept_confirm, leaving just
the click action on the delete link.