Brands: Manage deletion when dependent items exist.
Each Item now belongs_to a Brand, and we cannot delete
a brand that is relied on by an item. The existing
system test for deleting a brand fails, due to this
new change brought forth by Items - a fk violation is
raised.

Customize the association to restrict the deletion of
a Brand with an error, preventing the FK violation.

Test the association rule in the model test.
Modify the delete action in the controller.
Test the deletion of a brand without items, and also
with items.