Views: Normalize header links and table links.
Simplify ItemsController.
For #show, #edit, #update, and #destroy, there is no need
to depend on a work_order_id in the route nor load the
work order. We are given the id of an Item, and you can just
load that directly. If you do need the work order, you can
then just grab it from the loaded item.

References #74, #61, #62.