Repairs: Create objects necessary for form rendering upon validation failure.
When a repair is created with invalid attributes, the RepairsController#create
action will render the items/show view. That view now needs a new Discount,
a new Fee, a list of StandardFee objects, and a list of StandardDiscount
objects, in order to display all the forms.

Ensure that the #create action's else clause prepares these objects.

References #57 #58.