Price: Move and remove from production code.
Price is not a model, so move it into lib.
Stub a test class for Price. No class should exist w/o a unit test.

Right now, introducing Price this way is a premature abstraction,
and it exhibits a lack of clarity. Remove it from the fee calculation
in Item.