ItemsTest: Simplify adding item w/ different shipping.
We're about to introduce a bunch of additional items to the
shipped work order, which this test relies on. Rather than test for
adding an unshipped item to a work order that is shipping, do the
opposite: test for adding a shipped item on an unshipped work order.

This removes the complexity of the test as it is written, by using
a more vanilla work order without a bunch of items already in it.

References #107.