Item: Test inability to delete with fees, discounts, repairs independently.
The existing unit tests have a flaw, as the inability to delete
could be due to either having fees, discounts or repairs. In addition,
the skipped system tests beg for more isolated items to test.

Create three additional items: one with only a discount, one with
only a fee, and one with only a repair. Test the inability to
delete each of them, in both unit and system tests.

References #107.