WorkOrder: Update due date when customer is selected.
When creating a new work order without specifying a customer first,
the default due date is not known. As such, the model causes the
due date field to be the same as the in date.

When selecting a customer, update the due date based on that
customer's customer type and the expected turn-around for that
customer type.

References #170.