WorkOrder: Add 'open' scope method.
The dashboard needs to retrieve all non-invoiced work orders
to perform some calculations. Add a scope that loads work orders
that do not have an INVOICED ItemStatus.

Note that this will not retrieve work orders that have one
invoiced item and other non-invoiced items. We'll need to
seek clarification on this feature.

References #232.