DB: Remove intuit_accounts table.
Running the last migration causes a change to schema.rb. It seems the schema file
became out of sync from the migrations. Clean it up.

@theusaf With Rails, we have to be a little methodical with migrations, as when
we run them, it updates schema.rb. So if I change an existing migration without
undoing it first, debris in schema.rb can accumulate.

I do a rails db:rollback to undo the schema change, then modify the migration,
and re-run when I'm satisfied.