ItemsController: Fix regression due to missing customer.
When redirecting after a failed item addition to a work
order, the system test is failing due to the controller
not preparing all objects for re-rendering the work order
after the failed item addition.

This is a regression that may not have occurred before due
to some turbo magick.

Load the customer in the ItemsController#create failure path.